Déplacement des OneDimensionEnriched... dans OneDimensionEnrichedHansbo et suppression des fichiers
